+++ src/index.ts
@@ -1,8 +1,9 @@
 // SPDX-License-Identifier: CC0-1.0
 
-export { default, default as Exception } from "./Exception";
+export { default as Exception, default } from "./Exception";
 export { default as IllegalStateException } from "./IllegalStateException";
 export { default as InternalException } from "./InternalException";
 export { default as InvalidArgumentException } from "./InvalidArgumentException";
+export { default as MissingElementException } from "./MissingElementException";
 export { default as NoImplementationException } from "./NoImplementationException";
 export { default as OutOfRangeException } from "./OutOfRangeException";