+++ CHANGELOG.md
@@ -11,6 +11,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [**Keep a Changelog v1.0.0**](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [**Semantic Versioning v2.0.0**](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased ##
+
+### Added ###
+
+* pre-defined sub-exception: `MissingElementException`
+
+### Security ###
+
+* Updated dependencies
+
 ## [v1.3.0] - 2023-01-12 ##
 
 [v1.3.0]: https://github.com/mfederczuk/custom-js-exception/releases/tag/v1.3.0